+++ CHANGELOG.md
@@ -224,6 +224,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   definition file (#171).
 - Handle JSON decode error in `fatbuildrctl` in case of unexpected failure in
   `fatbuildrweb` (#146).
+- Report error instead of crashing in `fatbuildrctl` when unable to find format
+  corresponding to a given distribution with fatbuildrweb REST API (#172).
 - docs:
   - Add missing path parameter in REST API to retrieve artifact information.
   - Add missing optional `architectures` parameter in instances pipelines